chore: rename lib/base/finally.h => lib/base/FinalAction.h
