+++ src/lib/base/CMakeLists.txt
@@ -11,7 +11,7 @@ add_library(base STATIC
   EventQueue.h
   EventTypes.cpp
   EventTypes.h
-  finally.h
+  FinalAction.h
   FunctionEventJob.cpp
   FunctionEventJob.h
   FunctionJob.cpp
