+++ src/lib/net/SecureUtils.cpp
@@ -7,8 +7,8 @@
 
 #include "SecureUtils.h"
 
+#include "base/FinalAction.h"
 #include "base/String.h"
-#include "base/finally.h"
 #include "io/filesystem.h"
 
 #include <openssl/evp.h>
